+++ server/src/main/resources/schema.sql
@@ -60,7 +60,7 @@ create table if not exists `yozm-cafe`.un_viewed_cafe
 create table if not exists `yozm-cafe`.menu (
     id BIGINT NOT NULL AUTO_INCREMENT,
     cafe_id BIGINT,
-    priority BIGINT NOT NULL,
+    priority INT NOT NULL,
     name VARCHAR(255) NOT NULL,
     image_url VARCHAR(512),
     description VARCHAR(255),
